webgpu: export fromPixels as a module method
DEV
