+++ tfjs-backend-webgpu/src/kernels/FromPixels_utils/from_pixels_webgpu.ts
@@ -15,9 +15,9 @@
  * =============================================================================
  */
 
-import {computeDispatch, flatDispatchLayout} from '../webgpu_util';
+import {computeDispatch, flatDispatchLayout} from '../../webgpu_util';
 
-import {WebGPUProgram} from './webgpu_program';
+import {WebGPUProgram} from '../webgpu_program';
 
 export class FromPixelsProgram implements WebGPUProgram {
   outputShape: number[];
