+++ tfjs-backend-webgpu/src/register_all_kernels.ts
@@ -18,6 +18,7 @@ import {KernelConfig, registerKernel} from '@tensorflow/tfjs-core';
 
 import {divConfig} from './kernels/Div';
 import {fusedBatchNormConfig} from './kernels/FusedBatchNorm';
+import {fromPixelsConfig} from './kernels/FromPixels'
 import {mirrorPadConfig} from './kernels/MirrorPad';
 import {nonMaxSuppressionV3Config} from './kernels/NonMaxSuppressionV3';
 import {nonMaxSuppressionV5Config} from './kernels/NonMaxSuppressionV5';
@@ -33,6 +34,7 @@ const kernelConfigs: KernelConfig[] = [
   fusedBatchNormConfig,
   nonMaxSuppressionV3Config,
   nonMaxSuppressionV5Config,
+  fromPixelsConfig,
 ];
 
 for (const kernelConfig of kernelConfigs) {
